requirements.txt
alabaster==0.7.12
alembic==1.0.10
aniso8601==6.0.0
appdirs==1.4.3
astroid==2.2.5
atomicwrites==1.3.0
attrs==19.1.0
Babel==2.6.0
backcall==0.1.0
beautifulsoup4==4.7.1
black==18.9b0
blessings==1.7
bs4==0.0.1
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
Click==7.0
codecov==2.0.15
colorama==0.4.1
coverage==4.5.3
cssselect==1.0.3
decorator==4.4.0
docutils==0.14
entrypoints==0.3
fake-useragent==0.1.11
flake8==3.7.7
Flask==1.0.2
flask-marshmallow==0.10.1
Flask-Migrate==2.4.0
Flask-RESTful==0.3.7
flask-restful-swagger==0.20.1
flask-restplus==0.12.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
gevent==1.4.0
greenlet==0.4.15
gunicorn==19.9.0
idna==2.8
imagesize==1.1.0
inquirer==2.5.1
ipdb==0.12
ipython==7.5.0
ipython-genutils==0.2.0
isort==4.3.18
itsdangerous==1.1.0
jedi==0.13.3
Jinja2==2.10.1
jsonschema==3.0.1
lazy-object-proxy==1.4.0
lxml==4.3.3
Mako==1.0.10
MarkupSafe==1.1.1
marshmallow==2.19.2
marshmallow-sqlalchemy==0.16.3
mccabe==0.6.1
more-itertools==7.0.0
packaging==19.0
parse==1.12.0
parso==0.4.0
pbr==5.2.0
pickleshare==0.7.5
pluggy==0.11.0
prompt-toolkit==1.0.14
psycopg2==2.8.2
py==1.8.0
pycodestyle==2.5.0
pycparser==2.19
pyee==6.0.0
pyflakes==2.1.1
Pygments==2.4.0
PyInquirer==1.0.3
pylint==2.3.1
pyparsing==2.4.0
pyppeteer==0.0.25
pyquery==1.4.0
pyrsistent==0.15.2
pytest==4.5.0
pytest-cov==2.7.1
pytest-flask==0.14.0
python-dateutil==2.8.0
python-dotenv==0.10.2
python-editor==1.0.4
python-slugify==3.0.2
pytz==2019.1
readchar==2.0.1
redis==3.2.1
regex==2019.4.14
requests==2.21.0
requests-html==0.10.0
six==1.12.0
snowballstemmer==1.2.1
soupsieve==1.9.1
Sphinx==2.0.1
sphinxcontrib-apidoc==0.3.0
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
SQLAlchemy==1.3.3
text-unidecode==1.2
toml==0.10.0
tqdm==4.31.1
traitlets==4.3.2
typed-ast==1.3.5
urllib3==1.24.3
w3lib==1.20.0
wcwidth==0.1.7
websockets==7.0
Werkzeug==0.15.2
whaaaaat==0.5.2
wrapt==1.11.1
